+++ embassy-stm32/src/can/bxcan/mod.rs
@@ -1,28 +1,26 @@
+pub mod filter;
+mod registers;
+
 use core::future::poll_fn;
 use core::marker::PhantomData;
-use core::ops::{Deref, DerefMut};
 use core::task::Poll;
 
-pub mod bx;
-
-pub use bx::{filter, Data, ExtendedId, Fifo, Frame, Header, Id, StandardId};
 use embassy_hal_internal::{into_ref, PeripheralRef};
 use embassy_sync::blocking_mutex::raw::CriticalSectionRawMutex;
 use embassy_sync::channel::Channel;
 use embassy_sync::waitqueue::AtomicWaker;
+pub use embedded_can::{ExtendedId, Id, StandardId};
 
+use self::filter::MasterFilters;
+use self::registers::{Registers, RxFifo};
+pub use super::common::{BufferedCanReceiver, BufferedCanSender};
+use super::frame::{Envelope, Frame};
+use super::util;
+use crate::can::enums::{BusError, TryReadError};
 use crate::gpio::AFType;
 use crate::interrupt::typelevel::Interrupt;
 use crate::rcc::RccPeripheral;
 use crate::{interrupt, peripherals, Peripheral};
-
-pub mod enums;
-pub mod frame;
-pub mod util;
-pub use frame::Envelope;
-
-mod common;
-pub use self::common::{BufferedCanReceiver, BufferedCanSender};
 
 /// Interrupt handler.
 pub struct TxInterruptHandler<T: Instance> {
@@ -80,9 +78,72 @@ impl<T: Instance> interrupt::typelevel::Handler<T::SCEInterrupt> for SceInterrup
     }
 }
 
+/// Configuration proxy returned by [`Can::modify_config`].
+pub struct CanConfig<'a, T: Instance> {
+    can: PhantomData<&'a mut T>,
+}
+
+impl<T: Instance> CanConfig<'_, T> {
+    /// Configures the bit timings.
+    ///
+    /// You can use <http://www.bittiming.can-wiki.info/> to calculate the `btr` parameter. Enter
+    /// parameters as follows:
+    ///
+    /// - *Clock Rate*: The input clock speed to the CAN peripheral (*not* the CPU clock speed).
+    ///   This is the clock rate of the peripheral bus the CAN peripheral is attached to (eg. APB1).
+    /// - *Sample Point*: Should normally be left at the default value of 87.5%.
+    /// - *SJW*: Should normally be left at the default value of 1.
+    ///
+    /// Then copy the `CAN_BUS_TIME` register value from the table and pass it as the `btr`
+    /// parameter to this method.
+    pub fn set_bit_timing(self, bt: crate::can::util::NominalBitTiming) -> Self {
+        Registers(T::regs()).set_bit_timing(bt);
+        self
+    }
+
+    /// Configure the CAN bit rate.
+    ///
+    /// This is a helper that internally calls `set_bit_timing()`[Self::set_bit_timing].
+    pub fn set_bitrate(self, bitrate: u32) -> Self {
+        let bit_timing = util::calc_can_timings(T::frequency(), bitrate).unwrap();
+        self.set_bit_timing(bit_timing)
+    }
+
+    /// Enables or disables loopback mode: Internally connects the TX and RX
+    /// signals together.
+    pub fn set_loopback(self, enabled: bool) -> Self {
+        Registers(T::regs()).set_loopback(enabled);
+        self
+    }
+
+    /// Enables or disables silent mode: Disconnects the TX signal from the pin.
+    pub fn set_silent(self, enabled: bool) -> Self {
+        Registers(T::regs()).set_silent(enabled);
+        self
+    }
+
+    /// Enables or disables automatic retransmission of messages.
+    ///
+    /// If this is enabled, the CAN peripheral will automatically try to retransmit each frame
+    /// until it can be sent. Otherwise, it will try only once to send each frame.
+    ///
+    /// Automatic retransmission is enabled by default.
+    pub fn set_automatic_retransmit(self, enabled: bool) -> Self {
+        Registers(T::regs()).set_automatic_retransmit(enabled);
+        self
+    }
+}
+
+impl<T: Instance> Drop for CanConfig<'_, T> {
+    #[inline]
+    fn drop(&mut self) {
+        Registers(T::regs()).leave_init_mode();
+    }
+}
+
 /// CAN driver
 pub struct Can<'d, T: Instance> {
-    can: crate::can::bx::Can<BxcanInstance<'d, T>>,
+    peri: PeripheralRef<'d, T>,
 }
 
 /// Error returned by `try_write`
@@ -145,14 +206,25 @@ impl<'d, T: Instance> Can<'d, T> {
         rx.set_as_af(rx.af_num(), AFType::Input);
         tx.set_as_af(tx.af_num(), AFType::OutputPushPull);
 
-        let can = crate::can::bx::Can::builder(BxcanInstance(peri), T::regs()).leave_disabled();
-        Self { can }
+        Registers(T::regs()).leave_init_mode();
+
+        Self { peri }
     }
 
     /// Set CAN bit rate.
     pub fn set_bitrate(&mut self, bitrate: u32) {
         let bit_timing = util::calc_can_timings(T::frequency(), bitrate).unwrap();
-        self.can.modify_config().set_bit_timing(bit_timing).leave_disabled();
+        self.modify_config().set_bit_timing(bit_timing);
+    }
+
+    /// Configure bit timings and silent/loop-back mode.
+    ///
+    /// Calling this method will enter initialization mode. You must enable the peripheral
+    /// again afterwards with [`enable`](Self::enable).
+    pub fn modify_config(&mut self) -> CanConfig<'_, T> {
+        Registers(T::regs()).enter_init_mode();
+
+        CanConfig { can: PhantomData }
     }
 
     /// Enables the peripheral and synchronizes with the bus.
@@ -160,7 +232,7 @@ impl<'d, T: Instance> Can<'d, T> {
     /// This will wait for 11 consecutive recessive bits (bus idle state).
     /// Contrary to enable method from bxcan library, this will not freeze the executor while waiting.
     pub async fn enable(&mut self) {
-        while self.registers.enable_non_blocking().is_err() {
+        while Registers(T::regs()).enable_non_blocking().is_err() {
             // SCE interrupt is only generated for entering sleep mode, but not leaving.
             // Yield to allow other tasks to execute while can bus is initializing.
             embassy_futures::yield_now().await;
@@ -170,19 +242,19 @@ impl<'d, T: Instance> Can<'d, T> {
     /// Queues the message to be sent.
     ///
     /// If the TX queue is full, this will wait until there is space, therefore exerting backpressure.
-    pub async fn write(&mut self, frame: &Frame) -> crate::can::bx::TransmitStatus {
+    pub async fn write(&mut self, frame: &Frame) -> TransmitStatus {
         self.split().0.write(frame).await
     }
 
     /// Attempts to transmit a frame without blocking.
     ///
     /// Returns [Err(TryWriteError::Full)] if all transmit mailboxes are full.
-    pub fn try_write(&mut self, frame: &Frame) -> Result<crate::can::bx::TransmitStatus, TryWriteError> {
+    pub fn try_write(&mut self, frame: &Frame) -> Result<TransmitStatus, TryWriteError> {
         self.split().0.try_write(frame)
     }
 
     /// Waits for a specific transmit mailbox to become empty
-    pub async fn flush(&self, mb: crate::can::bx::Mailbox) {
+    pub async fn flush(&self, mb: Mailbox) {
         CanTx::<T>::flush_inner(mb).await
     }
 
@@ -194,6 +266,22 @@ impl<'d, T: Instance> Can<'d, T> {
     /// Waits until all of the transmit mailboxes become empty
     pub async fn flush_all(&self) {
         CanTx::<T>::flush_all_inner().await
+    }
+
+    /// Attempts to abort the sending of a frame that is pending in a mailbox.
+    ///
+    /// If there is no frame in the provided mailbox, or its transmission succeeds before it can be
+    /// aborted, this function has no effect and returns `false`.
+    ///
+    /// If there is a frame in the provided mailbox, and it is canceled successfully, this function
+    /// returns `true`.
+    pub fn abort(&mut self, mailbox: Mailbox) -> bool {
+        Registers(T::regs()).abort(mailbox)
+    }
+
+    /// Returns `true` if no frame is pending for transmission.
+    pub fn is_transmitter_idle(&self) -> bool {
+        Registers(T::regs()).is_idle()
     }
 
     /// Read a CAN frame.
@@ -221,8 +309,14 @@ impl<'d, T: Instance> Can<'d, T> {
     ///
     /// Useful for doing separate transmit/receive tasks.
     pub fn split<'c>(&'c mut self) -> (CanTx<'d, T>, CanRx<'d, T>) {
-        let (tx, rx) = self.can.split_by_ref();
-        (CanTx { tx }, CanRx { rx })
+        (
+            CanTx {
+                _peri: unsafe { self.peri.clone_unchecked() },
+            },
+            CanRx {
+                peri: unsafe { self.peri.clone_unchecked() },
+            },
+        )
     }
 
     /// Return a buffered instance of driver. User must supply Buffers
@@ -239,10 +333,13 @@ impl<'d, T: Instance> Can<'d, T> {
     }
 }
 
-impl<'d, T: Instance> AsMut<crate::can::bx::Can<BxcanInstance<'d, T>>> for Can<'d, T> {
-    /// Get mutable access to the lower-level driver from the `bxcan` crate.
-    fn as_mut(&mut self) -> &mut crate::can::bx::Can<BxcanInstance<'d, T>> {
-        &mut self.can
+impl<'d, T: FilterOwner> Can<'d, T> {
+    /// Accesses the filter banks owned by this CAN peripheral.
+    ///
+    /// To modify filters of a slave peripheral, `modify_filters` has to be called on the master
+    /// peripheral instead.
+    pub fn modify_filters(&mut self) -> MasterFilters<'_, T> {
+        unsafe { MasterFilters::new(T::regs()) }
     }
 }
 
@@ -288,17 +385,17 @@ impl<'d, T: Instance, const TX_BUF_SIZE: usize, const RX_BUF_SIZE: usize> Buffer
 
 /// CAN driver, transmit half.
 pub struct CanTx<'d, T: Instance> {
-    tx: crate::can::bx::Tx<BxcanInstance<'d, T>>,
+    _peri: PeripheralRef<'d, T>,
 }
 
 impl<'d, T: Instance> CanTx<'d, T> {
     /// Queues the message to be sent.
     ///
     /// If the TX queue is full, this will wait until there is space, therefore exerting backpressure.
-    pub async fn write(&mut self, frame: &Frame) -> crate::can::bx::TransmitStatus {
+    pub async fn write(&mut self, frame: &Frame) -> TransmitStatus {
         poll_fn(|cx| {
             T::state().tx_mode.register(cx.waker());
-            if let Ok(status) = self.tx.transmit(frame) {
+            if let Ok(status) = Registers(T::regs()).transmit(frame) {
                 return Poll::Ready(status);
             }
 
@@ -310,11 +407,11 @@ impl<'d, T: Instance> CanTx<'d, T> {
     /// Attempts to transmit a frame without blocking.
     ///
     /// Returns [Err(TryWriteError::Full)] if all transmit mailboxes are full.
-    pub fn try_write(&mut self, frame: &Frame) -> Result<crate::can::bx::TransmitStatus, TryWriteError> {
-        self.tx.transmit(frame).map_err(|_| TryWriteError::Full)
+    pub fn try_write(&mut self, frame: &Frame) -> Result<TransmitStatus, TryWriteError> {
+        Registers(T::regs()).transmit(frame).map_err(|_| TryWriteError::Full)
     }
 
-    async fn flush_inner(mb: crate::can::bx::Mailbox) {
+    async fn flush_inner(mb: Mailbox) {
         poll_fn(|cx| {
             T::state().tx_mode.register(cx.waker());
             if T::regs().tsr().read().tme(mb.index()) {
@@ -327,7 +424,7 @@ impl<'d, T: Instance> CanTx<'d, T> {
     }
 
     /// Waits for a specific transmit mailbox to become empty
-    pub async fn flush(&self, mb: crate::can::bx::Mailbox) {
+    pub async fn flush(&self, mb: Mailbox) {
         Self::flush_inner(mb).await
     }
 
@@ -336,9 +433,9 @@ impl<'d, T: Instance> CanTx<'d, T> {
             T::state().tx_mode.register(cx.waker());
 
             let tsr = T::regs().tsr().read();
-            if tsr.tme(crate::can::bx::Mailbox::Mailbox0.index())
-                || tsr.tme(crate::can::bx::Mailbox::Mailbox1.index())
-                || tsr.tme(crate::can::bx::Mailbox::Mailbox2.index())
+            if tsr.tme(Mailbox::Mailbox0.index())
+                || tsr.tme(Mailbox::Mailbox1.index())
+                || tsr.tme(Mailbox::Mailbox2.index())
             {
                 return Poll::Ready(());
             }
@@ -358,9 +455,9 @@ impl<'d, T: Instance> CanTx<'d, T> {
             T::state().tx_mode.register(cx.waker());
 
             let tsr = T::regs().tsr().read();
-            if tsr.tme(crate::can::bx::Mailbox::Mailbox0.index())
-                && tsr.tme(crate::can::bx::Mailbox::Mailbox1.index())
-                && tsr.tme(crate::can::bx::Mailbox::Mailbox2.index())
+            if tsr.tme(Mailbox::Mailbox0.index())
+                && tsr.tme(Mailbox::Mailbox1.index())
+                && tsr.tme(Mailbox::Mailbox2.index())
             {
                 return Poll::Ready(());
             }
@@ -375,12 +472,28 @@ impl<'d, T: Instance> CanTx<'d, T> {
         Self::flush_all_inner().await
     }
 
+    /// Attempts to abort the sending of a frame that is pending in a mailbox.
+    ///
+    /// If there is no frame in the provided mailbox, or its transmission succeeds before it can be
+    /// aborted, this function has no effect and returns `false`.
+    ///
+    /// If there is a frame in the provided mailbox, and it is canceled successfully, this function
+    /// returns `true`.
+    pub fn abort(&mut self, mailbox: Mailbox) -> bool {
+        Registers(T::regs()).abort(mailbox)
+    }
+
+    /// Returns `true` if no frame is pending for transmission.
+    pub fn is_idle(&self) -> bool {
+        Registers(T::regs()).is_idle()
+    }
+
     /// Return a buffered instance of driver. User must supply Buffers
     pub fn buffered<const TX_BUF_SIZE: usize>(
         self,
         txb: &'static mut TxBuf<TX_BUF_SIZE>,
     ) -> BufferedCanTx<'d, T, TX_BUF_SIZE> {
-        BufferedCanTx::new(self.tx, txb)
+        BufferedCanTx::new(self, txb)
     }
 }
 
@@ -389,19 +502,19 @@ pub type TxBuf<const BUF_SIZE: usize> = Channel<CriticalSectionRawMutex, Frame, 
 
 /// Buffered CAN driver, transmit half.
 pub struct BufferedCanTx<'d, T: Instance, const TX_BUF_SIZE: usize> {
-    _tx: crate::can::bx::Tx<BxcanInstance<'d, T>>,
+    _tx: CanTx<'d, T>,
     tx_buf: &'static TxBuf<TX_BUF_SIZE>,
 }
 
 impl<'d, T: Instance, const TX_BUF_SIZE: usize> BufferedCanTx<'d, T, TX_BUF_SIZE> {
-    fn new(_tx: crate::can::bx::Tx<BxcanInstance<'d, T>>, tx_buf: &'static TxBuf<TX_BUF_SIZE>) -> Self {
+    fn new(_tx: CanTx<'d, T>, tx_buf: &'static TxBuf<TX_BUF_SIZE>) -> Self {
         Self { _tx, tx_buf }.setup()
     }
 
     fn setup(self) -> Self {
         // We don't want interrupts being processed while we change modes.
         critical_section::with(|_| unsafe {
-            let tx_inner = self::common::ClassicBufferedTxInner {
+            let tx_inner = super::common::ClassicBufferedTxInner {
                 tx_receiver: self.tx_buf.receiver().into(),
             };
             T::mut_state().tx_mode = TxMode::Buffered(tx_inner);
@@ -435,7 +548,7 @@ impl<'d, T: Instance, const TX_BUF_SIZE: usize> Drop for BufferedCanTx<'d, T, TX
 /// CAN driver, receive half.
 #[allow(dead_code)]
 pub struct CanRx<'d, T: Instance> {
-    rx: crate::can::bx::Rx<BxcanInstance<'d, T>>,
+    peri: PeripheralRef<'d, T>,
 }
 
 impl<'d, T: Instance> CanRx<'d, T> {
@@ -465,7 +578,7 @@ impl<'d, T: Instance> CanRx<'d, T> {
         self,
         rxb: &'static mut RxBuf<RX_BUF_SIZE>,
     ) -> BufferedCanRx<'d, T, RX_BUF_SIZE> {
-        BufferedCanRx::new(self.rx, rxb)
+        BufferedCanRx::new(self, rxb)
     }
 }
 
@@ -474,19 +587,19 @@ pub type RxBuf<const BUF_SIZE: usize> = Channel<CriticalSectionRawMutex, Result<
 
 /// CAN driver, receive half in Buffered mode.
 pub struct BufferedCanRx<'d, T: Instance, const RX_BUF_SIZE: usize> {
-    _rx: crate::can::bx::Rx<BxcanInstance<'d, T>>,
+    _rx: CanRx<'d, T>,
     rx_buf: &'static RxBuf<RX_BUF_SIZE>,
 }
 
 impl<'d, T: Instance, const RX_BUF_SIZE: usize> BufferedCanRx<'d, T, RX_BUF_SIZE> {
-    fn new(_rx: crate::can::bx::Rx<BxcanInstance<'d, T>>, rx_buf: &'static RxBuf<RX_BUF_SIZE>) -> Self {
+    fn new(_rx: CanRx<'d, T>, rx_buf: &'static RxBuf<RX_BUF_SIZE>) -> Self {
         BufferedCanRx { _rx, rx_buf }.setup()
     }
 
     fn setup(self) -> Self {
         // We don't want interrupts being processed while we change modes.
         critical_section::with(|_| unsafe {
-            let rx_inner = self::common::ClassicBufferedRxInner {
+            let rx_inner = super::common::ClassicBufferedRxInner {
                 rx_sender: self.rx_buf.sender().into(),
             };
             T::mut_state().rx_mode = RxMode::Buffered(rx_inner);
@@ -511,8 +624,7 @@ impl<'d, T: Instance, const RX_BUF_SIZE: usize> BufferedCanRx<'d, T, RX_BUF_SIZE
                         Err(e) => Err(TryReadError::BusError(e)),
                     }
                 } else {
-                    let registers = crate::can::bx::Registers { canregs: T::regs() };
-                    if let Some(err) = registers.curr_error() {
+                    if let Some(err) = Registers(T::regs()).curr_error() {
                         return Err(TryReadError::BusError(err));
                     } else {
                         Err(TryReadError::Empty)
@@ -544,8 +656,6 @@ impl<'d, T: Instance, const RX_BUF_SIZE: usize> Drop for BufferedCanRx<'d, T, RX
     }
 }
 
-use crate::can::bx::RxFifo;
-
 impl<'d, T: Instance> Drop for Can<'d, T> {
     fn drop(&mut self) {
         // Cannot call `free()` because it moves the instance.
@@ -555,35 +665,62 @@ impl<'d, T: Instance> Drop for Can<'d, T> {
     }
 }
 
-impl<'d, T: Instance> Deref for Can<'d, T> {
-    type Target = crate::can::bx::Can<BxcanInstance<'d, T>>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.can
-    }
+/// Identifies one of the two receive FIFOs.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum Fifo {
+    /// First receive FIFO
+    Fifo0 = 0,
+    /// Second receive FIFO
+    Fifo1 = 1,
 }
 
-impl<'d, T: Instance> DerefMut for Can<'d, T> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.can
-    }
+/// Identifies one of the three transmit mailboxes.
+#[derive(Debug, Copy, Clone, Ord, PartialOrd, Eq, PartialEq)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum Mailbox {
+    /// Transmit mailbox 0
+    Mailbox0 = 0,
+    /// Transmit mailbox 1
+    Mailbox1 = 1,
+    /// Transmit mailbox 2
+    Mailbox2 = 2,
 }
 
-use crate::can::enums::{BusError, TryReadError};
+/// Contains information about a frame enqueued for transmission via [`Can::transmit`] or
+/// [`Tx::transmit`].
+pub struct TransmitStatus {
+    dequeued_frame: Option<Frame>,
+    mailbox: Mailbox,
+}
+
+impl TransmitStatus {
+    /// Returns the lower-priority frame that was dequeued to make space for the new frame.
+    #[inline]
+    pub fn dequeued_frame(&self) -> Option<&Frame> {
+        self.dequeued_frame.as_ref()
+    }
+
+    /// Returns the [`Mailbox`] the frame was enqueued in.
+    #[inline]
+    pub fn mailbox(&self) -> Mailbox {
+        self.mailbox
+    }
+}
 
 pub(crate) enum RxMode {
     NonBuffered(AtomicWaker),
-    Buffered(crate::can::_version::common::ClassicBufferedRxInner),
+    Buffered(super::common::ClassicBufferedRxInner),
 }
 
 impl RxMode {
-    pub fn on_interrupt<T: Instance>(&self, fifo: crate::can::_version::bx::RxFifo) {
+    pub fn on_interrupt<T: Instance>(&self, fifo: RxFifo) {
         match self {
             Self::NonBuffered(waker) => {
                 // Disable interrupts until read
                 let fifo_idx = match fifo {
-                    crate::can::_version::bx::RxFifo::Fifo0 => 0usize,
-                    crate::can::_version::bx::RxFifo::Fifo1 => 1usize,
+                    RxFifo::Fifo0 => 0usize,
+                    RxFifo::Fifo1 => 1usize,
                 };
                 T::regs().ier().write(|w| {
                     w.set_fmpie(fifo_idx, false);
@@ -591,10 +728,8 @@ impl RxMode {
                 waker.wake();
             }
             Self::Buffered(buf) => {
-                let regsisters = crate::can::bx::Registers { canregs: T::regs() };
-
                 loop {
-                    match regsisters.receive_fifo(fifo) {
+                    match Registers(T::regs()).receive_fifo(fifo) {
                         Some(envelope) => {
                             // NOTE: consensus was reached that if rx_queue is full, packets should be dropped
                             let _ = buf.rx_sender.try_send(Ok(envelope));
@@ -628,13 +763,13 @@ impl RxMode {
     pub fn try_read<T: Instance>(&self) -> Result<Envelope, TryReadError> {
         match self {
             Self::NonBuffered(_) => {
-                let registers = crate::can::bx::Registers { canregs: T::regs() };
-                if let Some(msg) = registers.receive_fifo(super::bx::RxFifo::Fifo0) {
+                let registers = Registers(T::regs());
+                if let Some(msg) = registers.receive_fifo(RxFifo::Fifo0) {
                     T::regs().ier().write(|w| {
                         w.set_fmpie(0, true);
                     });
                     Ok(msg)
-                } else if let Some(msg) = registers.receive_fifo(super::bx::RxFifo::Fifo1) {
+                } else if let Some(msg) = registers.receive_fifo(RxFifo::Fifo1) {
                     T::regs().ier().write(|w| {
                         w.set_fmpie(1, true);
                     });
@@ -655,8 +790,7 @@ impl RxMode {
             Self::NonBuffered(waker) => {
                 poll_fn(|cx| {
                     waker.register(cx.waker());
-                    let registers = crate::can::bx::Registers { canregs: T::regs() };
-                    if registers.receive_frame_available() {
+                    if Registers(T::regs()).receive_frame_available() {
                         Poll::Ready(())
                     } else {
                         Poll::Pending
@@ -673,15 +807,13 @@ impl RxMode {
 
 enum TxMode {
     NonBuffered(AtomicWaker),
-    Buffered(self::common::ClassicBufferedTxInner),
+    Buffered(super::common::ClassicBufferedTxInner),
 }
 
 impl TxMode {
     pub fn buffer_free<T: Instance>(&self) -> bool {
         let tsr = T::regs().tsr().read();
-        tsr.tme(crate::can::bx::Mailbox::Mailbox0.index())
-            || tsr.tme(crate::can::bx::Mailbox::Mailbox1.index())
-            || tsr.tme(crate::can::bx::Mailbox::Mailbox2.index())
+        tsr.tme(Mailbox::Mailbox0.index()) || tsr.tme(Mailbox::Mailbox1.index()) || tsr.tme(Mailbox::Mailbox2.index())
     }
     pub fn on_interrupt<T: Instance>(&self) {
         match &T::state().tx_mode {
@@ -690,8 +822,7 @@ impl TxMode {
                 while self.buffer_free::<T>() {
                     match buf.tx_receiver.try_receive() {
                         Ok(frame) => {
-                            let mut registers = crate::can::bx::Registers { canregs: T::regs() };
-                            _ = registers.transmit(&frame);
+                            _ = Registers(T::regs()).transmit(&frame);
                         }
                         Err(_) => {
                             break;
@@ -738,7 +869,7 @@ trait SealedInstance {
 
 /// CAN instance trait.
 #[allow(private_bounds)]
-pub trait Instance: SealedInstance + RccPeripheral + 'static {
+pub trait Instance: Peripheral<P = Self> + SealedInstance + RccPeripheral + 'static {
     /// TX interrupt for this instance.
     type TXInterrupt: crate::interrupt::typelevel::Interrupt;
     /// RX0 interrupt for this instance.
@@ -749,10 +880,35 @@ pub trait Instance: SealedInstance + RccPeripheral + 'static {
     type SCEInterrupt: crate::interrupt::typelevel::Interrupt;
 }
 
-/// BXCAN instance newtype.
-pub struct BxcanInstance<'a, T>(PeripheralRef<'a, T>);
+/// A bxCAN instance that owns filter banks.
+///
+/// In master-slave-instance setups, only the master instance owns the filter banks, and needs to
+/// split some of them off for use by the slave instance. In that case, the master instance should
+/// implement [`FilterOwner`] and [`MasterInstance`], while the slave instance should only implement
+/// [`Instance`].
+///
+/// In single-instance configurations, the instance owns all filter banks and they can not be split
+/// off. In that case, the instance should implement [`Instance`] and [`FilterOwner`].
+///
+/// # Safety
+///
+/// This trait must only be implemented if the instance does, in fact, own its associated filter
+/// banks, and `NUM_FILTER_BANKS` must be correct.
+pub unsafe trait FilterOwner: Instance {
+    /// The total number of filter banks available to the instance.
+    ///
+    /// This is usually either 14 or 28, and should be specified in the chip's reference manual or datasheet.
+    const NUM_FILTER_BANKS: u8;
+}
 
-unsafe impl<'d, T: Instance> crate::can::bx::Instance for BxcanInstance<'d, T> {}
+/// A bxCAN master instance that shares filter banks with a slave instance.
+///
+/// In master-slave-instance setups, this trait should be implemented for the master instance.
+///
+/// # Safety
+///
+/// This trait must only be implemented when there is actually an associated slave instance.
+pub unsafe trait MasterInstance: FilterOwner {}
 
 foreach_peripheral!(
     (can, $inst:ident) => {
@@ -782,7 +938,7 @@ foreach_peripheral!(
 
 foreach_peripheral!(
     (can, CAN) => {
-        unsafe impl<'d> crate::can::bx::FilterOwner for BxcanInstance<'d, peripherals::CAN> {
+        unsafe impl FilterOwner for peripherals::CAN {
             const NUM_FILTER_BANKS: u8 = 14;
         }
     };
@@ -797,19 +953,19 @@ foreach_peripheral!(
             ))] {
                 // Most L4 devices and some F7 devices use the name "CAN1"
                 // even if there is no "CAN2" peripheral.
-                unsafe impl<'d> crate::can::bx::FilterOwner for BxcanInstance<'d, peripherals::CAN1> {
+                unsafe impl FilterOwner for peripherals::CAN1 {
                     const NUM_FILTER_BANKS: u8 = 14;
                 }
             } else {
-                unsafe impl<'d> crate::can::bx::FilterOwner for BxcanInstance<'d, peripherals::CAN1> {
+                unsafe impl FilterOwner for peripherals::CAN1 {
                     const NUM_FILTER_BANKS: u8 = 28;
                 }
-                unsafe impl<'d> crate::can::bx::MasterInstance for BxcanInstance<'d, peripherals::CAN1> {}
+                unsafe impl MasterInstance for peripherals::CAN1 {}
             }
         }
     };
     (can, CAN3) => {
-        unsafe impl<'d> crate::can::bx::FilterOwner for BxcanInstance<'d, peripherals::CAN3> {
+        unsafe impl FilterOwner for peripherals::CAN3 {
             const NUM_FILTER_BANKS: u8 = 14;
         }
     };
@@ -822,12 +978,12 @@ trait Index {
     fn index(&self) -> usize;
 }
 
-impl Index for crate::can::bx::Mailbox {
+impl Index for Mailbox {
     fn index(&self) -> usize {
         match self {
-            crate::can::bx::Mailbox::Mailbox0 => 0,
-            crate::can::bx::Mailbox::Mailbox1 => 1,
-            crate::can::bx::Mailbox::Mailbox2 => 2,
+            Mailbox::Mailbox0 => 0,
+            Mailbox::Mailbox1 => 1,
+            Mailbox::Mailbox2 => 2,
         }
     }
 }
